+++ R/docinfo.R
@@ -98,13 +98,6 @@ entry_pdftk <- function(key, value) {
       paste("InfoValue:", value))
 }
 
-get_pdftk_metadata <- function(filename) {
-    cmd <- pdftk()
-    filename <- shQuote(normalizePath(filename, mustWork = TRUE))
-    args <- c(filename, "dump_data_utf8")
-    system2(cmd, args, stdout=TRUE)
-}
-
 #' @rdname docinfo
 #' @export
 get_docinfo <- function(filename) {