+++ NEWS.md
@@ -34,6 +34,7 @@ Initial features
   1. `n_pages_qpdf()` which wraps `qpdf::pdf_length()`
   2. `n_pages_pdftk()` which wraps `pdftk` command-line tool
   3. `n_pages_gs()` which wraps `ghostscript` command-line tool
+  4. `n_pages_exiftool()` which wraps `exiftool` command-line tool
 
 * The following functions detect support for various higher-level features:
 
@@ -46,5 +47,6 @@ Initial features
 * The following functions detect support for various command-line tools needed for some
   lower-level helper functions:
 
+  * `supports_exiftool()` detects support for `exiftool`
   * `supports_gs()` detects support for `ghostscript`
   * `supports_pdftk()` detects support for `pdftk`